# Optimize file processing 