--- a/server/src/controller/fileController.ts
+++ b/server/src/controller/fileController.ts
@@ -33,7 +33,6 @@ export const fileController = {
 		}
 
 		// File uploaded successfully
-		console.log("File uploaded successfully.");
 		next();
 		return;
 	},