--- a/server/src/utils/csv2GeojsonUtils.ts
+++ b/server/src/utils/csv2GeojsonUtils.ts
@@ -1,9 +1,6 @@
-import fs from "fs";
-import fsp from "fs/promises";
-import path from "path";
 import csvtojson from "csvtojson";
 
-async function convertCSVToGeoJSON(csvFilePath: string, filename: string) {
+async function convertCSVToGeoJSON(csvFilePath: string, _filename: string) {
 	try {
 		const csvToJson: any[] = await csvtojson().fromFile(csvFilePath);
 		const features: any[] = csvToJson
@@ -26,13 +23,10 @@ async function convertCSVToGeoJSON(csvFilePath: string, filename: string) {
 			features: features,
 		};
 
-		try {
-			return JSON.stringify(geoJSON, null, 2);
-		} catch (error) {
-			console.error("Failed to write GeoJSON file:", error);
-		}
+		return JSON.stringify(geoJSON, null, 2);
 	} catch (error) {
 		console.error("Failed to convert CSV to GeoJSON:", error);
+		return null;
 	}
 }