#### Things done:
- Use the memory more in favor of file read/write. 
- The files that are temporarily saved now are the `original CSV` and the `totalSamples final file`.
### File processing Times

##### Wheats File Optimization Cycle:
Initial Time: `4.317 seconds`
Optimized Time: `1.174 seconds`
Improvement: Reduction of `3.143 seconds`
Percentage Reduction: `72.79%`

##### Canola Optimization Cycle:
Initial Time: `4.598 seconds`
Optimized Time: `0.695 seconds`
Improvement: Reduction of `3.904 seconds`
Percentage Reduction: `84.9%`

#### Things to consider:
- Optimize even further to reduce file write/reads.